Optimize default depth grid, fix sysconfig read (#44)
The default depth grid is now determined based on the distribution of
the pressure readings which is especially important for moorings
experiencing strong knockdowns.

This commit also fixes a bug in reading the sysconfig which in turn is
necessary for reading the instrument orientation. The configuration was
read in the middle of the time series, now it is read in the middle of
the time series below surface only.